Split ecosystem implementations into own files
