+++ osv/ecosystems/rubygems_test.py
@@ -1,0 +1,34 @@
+# Copyright 2021 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""RubyGems ecosystem helper tests."""
+
+import unittest
+
+from .. import ecosystems
+
+class RubyGemsEcosystemTest(unittest.TestCase):
+  """RubyGems ecosystem helper tests."""
+  
+  def test_next_version(self):
+    """Test next_version."""
+    ecosystem = ecosystems.get('RubyGems')
+    self.assertEqual('0.8.0', ecosystem.next_version('rails', '0'))
+    self.assertEqual('0.9.5', ecosystem.next_version('rails', '0.9.4.1'))
+    self.assertEqual('2.3.8.pre1', ecosystem.next_version('rails', '2.3.7'))
+    self.assertEqual('4.0.0.rc1',
+                     ecosystem.next_version('rails', '4.0.0.beta1'))
+    self.assertEqual('5.0.0.racecar1',
+                     ecosystem.next_version('rails', '5.0.0.beta4'))
+    with self.assertRaises(ecosystems.EnumerateError):
+      ecosystem.next_version('doesnotexist123456', '1')
