+++ osv/ecosystems/semver.py
@@ -1,0 +1,56 @@
+# Copyright 2021 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""SemVer ecosystem helper."""
+
+from .helper_base import Ecosystem
+from .. import semver_index
+
+
+class SemverEcosystem(Ecosystem):
+  """Generic semver ecosystem helpers."""
+
+  def sort_key(self, version):
+    """Sort key."""
+    return semver_index.parse(version)
+
+  def enumerate_versions(self,
+                         package,
+                         introduced,
+                         fixed=None,
+                         last_affected=None,
+                         limits=None):
+    """Enumerate versions (no-op)."""
+    del package
+    del introduced
+    del fixed
+    del limits
+
+  def next_version(self, package, version):
+    """Get the next version after the given version."""
+    del package  # Unused.
+    parsed_version = semver_index.parse(version)
+    if parsed_version.prerelease:
+      return version + '.0'
+
+    return str(parsed_version.bump_patch()) + '-0'
+
+  @property
+  def is_semver(self):
+    return True
+
+
+Crates = SemverEcosystem
+Go = SemverEcosystem
+NPM = SemverEcosystem
+Hex = SemverEcosystem
