+++ run_tests.sh
@@ -7,9 +7,14 @@ test -z "$DEPSDEV_API_KEY" && echo "FYI no API key for deps.dev, tests needing i
 set -x
 python3 -m pipenv sync
 python3 -m pipenv run python -m unittest osv.bug_test
-python3 -m pipenv run python -m unittest osv.ecosystems.ecosystems_test
-python3 -m pipenv run python -m unittest osv.ecosystems.maven.version_test
+python3 -m pipenv run python -m unittest osv.ecosystems.alpine_test
+python3 -m pipenv run python -m unittest osv.ecosystems.debian_test
+python3 -m pipenv run python -m unittest osv.ecosystems.maven_test
 python3 -m pipenv run python -m unittest osv.ecosystems.nuget_test
+python3 -m pipenv run python -m unittest osv.ecosystems.packagist_test
+python3 -m pipenv run python -m unittest osv.ecosystems.pypi_test
+python3 -m pipenv run python -m unittest osv.ecosystems.rubygems_test
+python3 -m pipenv run python -m unittest osv.ecosystems.semver_test
 python3 -m pipenv run python -m unittest osv.purl_helpers_test
 python3 -m pipenv run python -m unittest osv.request_helper_test
 python3 -m pipenv run python -m unittest osv.semver_index_test
